Update _Application.xml
Attempting to address a discrepancy between the method signature at the top and the explanation of default boolean values for the optional parameters. In this proposed change I've updated the parameter details to match the signature - however, in recent work with Microsoft.Office.Interop.Word.dll I have noticed discrepancies where one version will result in IL showing the defaults as true and the other as false - so honestly not 100% sure whether the method signature is correct or the parameter documentation.